+++ server/src/application/clips/use-case/query-handler/get-user-client-variant-clips-query-handler.ts
@@ -1,0 +1,5 @@
+import { GetUserClientVariantClipsQuery } from './query/get-user-client-variant-clips-query'
+
+const getUserClientVariantClipsQueryHandler = (
+  query: GetUserClientVariantClipsQuery
+) => true
