+++ server/src/application/clips/use-case/query-handler/query/get-user-client-variant-clips-query.ts
@@ -1,0 +1,4 @@
+export type GetUserClientVariantClipsQuery = {
+  clientId: string
+  locale: string
+}
