+++ server/src/application/reports/use-case/command-handler/create-report-command-handler.ts
@@ -1,4 +1,4 @@
-import { insertReportIntoDb } from '../../repository/reports-repository'
+import { insertReportIntoDb } from '../../../repository/reports-repository'
 import { CreateReportCommand } from './command/create-report-command'
 
 export const createReportCommandHandler = (command: CreateReportCommand) => {
