+++ server/src/application/sentences/use-case/command-handler/add-bulk-sentences-command-handler.ts
@@ -1,11 +1,11 @@
 import { pipe } from 'fp-ts/lib/function'
 import * as O from 'fp-ts/Option'
-import { InsertBulkSentences } from '../../repository/sentences-repository'
+import { InsertBulkSentences } from '../../../repository/sentences-repository'
 import { taskEither as TE } from 'fp-ts'
 import { AddBulkSentencesCommand } from './command/add-bulk-sentences-command'
 import { ReadTsvIntoMemory } from '../../../../infrastructure/parser/tsvParser'
-import { FetchUserClientIdByEmail } from '../../repository/user-repository'
-import { FetchSentenceDomains } from '../../repository/domain-repository'
+import { FetchUserClientIdByEmail } from '../../../repository/user-repository'
+import { FetchSentenceDomains } from '../../../repository/domain-repository'
 import { SentenceDomainDescription } from 'common'
 import { SentenceSubmission } from '../../../types/sentence-submission'
