+++ server/src/application/sentences/use-case/command-handler/add-sentence-vote-command-handler.ts
@@ -1,4 +1,4 @@
-import { insertSentenceVoteIntoDb } from '../../repository/sentences-repository'
+import { insertSentenceVoteIntoDb } from '../../../repository/sentences-repository'
 import { AddSentenceVoteCommand } from './command/add-sentence-vote-command'
 
 export const addSentenceVoteCommandHandler = (
