+++ server/src/application/sentences/use-case/query-handler/get-sentences-for-review-query-handler.ts
@@ -1,10 +1,10 @@
 import { pipe } from 'fp-ts/lib/function'
 import { option as O, task as T, taskOption as TO } from 'fp-ts'
 import * as TE from 'fp-ts/TaskEither'
-import { FindSentencesForReview } from '../../repository/sentences-repository'
+import { FindSentencesForReview } from '../../../repository/sentences-repository'
 import { GetSentencesForReviewQuery } from './query/get-sentences-for-review-query'
 import { UnvalidatedSentenceDto } from '../../types/unvalidatedSentenceDto'
-import { FetchUserClientVariants } from '../../../variants/repository/user-client-variants-repository'
+import { FetchUserClientVariants } from '../../../repository/user-client-variants-repository'
 
 export const GetSentencesForReviewQueryHandler =
   (fetchUserClientVariants: FetchUserClientVariants) =>
